# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Public and admin API for the Chatkhil RCY community website
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/auth/me:
    get:
      operationId: getAdminSession
      tags: [admin]
      summary: Get the current admin user
      responses:
        "200":
          description: Current admin session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not signed in
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [public]
      summary: Request a short-lived upload URL for a photo or media file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadRequest"
      responses:
        "200":
          description: Upload destination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadRequestResponse"
components:
  schemas:
...
    AdminUser:
      type: object
      properties:
        email: { type: string, format: email }
        role: { type: string, enum: [super_admin, editor] }
      required: [email, role]
...
    UploadRequest:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        size: { type: integer, minimum: 1 }
        contentType: { type: string, minLength: 1 }
      required: [name, size, contentType]
    UploadRequestResponse:
      type: object
      properties:
        uploadURL: { type: string, format: uri }
        objectPath: { type: string }
      required: [uploadURL, objectPath]