# .migration-backup/lib/api-zod/src/generated/types/uploadRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public and admin API for the Chatkhil RCY community website
 * OpenAPI spec version: 0.1.0
 */

export interface UploadRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}