# .migration-backup/lib/api-zod/src/generated/types/uploadRequestResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public and admin API for the Chatkhil RCY community website
 * OpenAPI spec version: 0.1.0
 */

export interface UploadRequestResponse {
  uploadURL: string;
  objectPath: string;
}